Unified strategy into one single file.
We can simply have one file to load all the strategies. So, they are all included in a file with the entity that uses it.
