--- app/src/main/java/ayds/newyork/songinfo/home/model/patterns/strategy/songs/SpotifySongReleaseDateByDefaultStrategy.kt
@@ -1,9 +1,0 @@
-package ayds.newyork.songinfo.home.model.patterns.strategy.songs
-
-class SpotifySongReleaseDateByDefaultStrategy: SpotifySongReleaseDateStrategy {
-
-    override fun printReleaseDate(releaseDate: String): String {
-        return ""
-    }
-
-}
