--- app/src/main/java/ayds/newyork/songinfo/home/model/patterns/strategy/songs/SpotifySongReleaseDateByYearStrategy.kt
@@ -1,9 +1,0 @@
-package ayds.newyork.songinfo.home.model.patterns.strategy.songs
-
-class SpotifySongReleaseDateByYearStrategy: SpotifySongReleaseDateStrategy {
-
-    override fun printReleaseDate(releaseDate: String): String {
-        return releaseDate.split("-").first()
-    }
-
-}
