--- app/src/main/java/ayds/newyork/songinfo/home/model/patterns/strategy/songs/SpotifySongReleaseDateStrategy.kt
@@ -1,7 +1,0 @@
-package ayds.newyork.songinfo.home.model.patterns.strategy.songs
-
-interface SpotifySongReleaseDateStrategy {
-
-    fun printReleaseDate(releaseDate: String): String
-
-}
